requirements.txt
aiohttp>=3.7.4.post0
asn1crypto==0.24.0
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
backports.zoneinfo==0.2.1
binance==0.3
blessed==1.17.6
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
cycler==0.10.0
dateparser==1.1.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.1
kiwisolver==1.3.1
matplotlib==3.3.4
multidict==5.2.0
nginxparser_eb==0.0.7
numpy==1.19.5
pandas==1.1.5
Pillow==8.3.2
pyparsing==2.4.7
python-binance==1.0.15
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
readchar==2.0.1
regex==2021.10.8
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.25
ta==0.7.0
typing-extensions==3.10.0.2
tzdata==2021.2.post0
tzlocal==3.0
ujson==4.2.0
urllib3==1.26.7
websockets==9.1
yarl==1.7.0
yfinance==0.1.63
zipp==3.6.0
